+++ src/tests/tests.html
@@ -66,8 +66,8 @@
 	<script type="text/javascript" src="http/clipperCachedHttp_tests.js"></script>
 	<script type="text/javascript" src="http/http_tests.js"></script>
 
-	<script type="text/javascript" src="localization/localization.js"></script>
-	<script type="text/javascript" src="localization/rtl.js"></script>
+	<script type="text/javascript" src="localization/localization_tests.js"></script>
+	<script type="text/javascript" src="localization/rtl_tests.js"></script>
 
 	<script type="text/javascript" src="logging/communicatorLoggerPure_tests.js"></script>
 	<script type="text/javascript" src="logging/consoleLoggerShell_tests.js"></script>
@@ -83,12 +83,11 @@
 	<script type="text/javascript" src="versioning/version_tests.js"></script>
 	<script type="text/javascript" src="versioning/changeLogHelper_tests.js"></script>
 
-	<script type="text/javascript" src="constants_tests.js"></script>
+	<script type="text/javascript" src="arrayUtils_tests.js"></script>
 	<script type="text/javascript" src="experiments_tests.js"></script>
-	<script type="text/javascript" src="rtl_tests.js"></script>
 	<script type="text/javascript" src="settings_tests.js"></script>
-	<script type="text/javascript" src="utils_tests.js"></script>
 	<script type="text/javascript" src="stringUtils_tests.js"></script>
+	<script type="text/javascript" src="utils_tests.js"></script>
 </head>
 <body>
 	<div id="qunit"></div>
